pyproject.toml
[tool.poetry]
name = "mywellnezz"
version = "1.2.1"
description = ""
authors = ["AeonDave"]
readme = "README.md"
packages = [
    { include = "mywellnezz" },
    { include = "app", from = "mywellnezz" },
    { include = "models", from = "mywellnezz" },
    { include = "modules", from = "mywellnezz" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
prettytable = "^3"
colorama = "^0.4"
psutil = "^6"
aioconsole = "^0.8"
pillow = "^10"
pwinput = "^1"
python-dateutil = "^2"
loguru = "^0.7"
pyinstaller = "^6"
mouse = "^0.7"
httpx = "^0.27"

[tool.poetry.scripts]
mywellnezz = "mywellnezz.main:start"
build = "pyinstaller:install"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
